structures: refactoring -- change long weight -> double weight for WeightedGraph

+++ data-structures/src/main/java/com/gitlab/sszuev/graphs/ModifiableWeightedGraph.java
@@ -10,18 +10,18 @@ public interface ModifiableWeightedGraph<X> extends WeightedGraph<X> {
      *
      * @param left   {@link X} - a payload for the left vertex, not {@code null}
      * @param right  {@link X} - a payload for the right vertex, not {@code null}
-     * @param weight {@code long} a weight of edge
+     * @param weight {@code double} a weight of edge
      * @return {@link Edge}
      */
-    Edge<X> add(X left, X right, long weight);
+    Edge<X> add(X left, X right, double weight);
 
     /**
      * @param left   {@link X} - a payload for the left vertex, not {@code null}
      * @param right  {@link X} - a payload for the right vertex, not {@code null}
-     * @param weight {@code long} a weight of edge
+     * @param weight {@code double} a weight of edge
      * @return a <b>this</b> graph to allow cascading calls
      */
-    default ModifiableWeightedGraph<X> addEdge(X left, X right, long weight) {
+    default ModifiableWeightedGraph<X> addEdge(X left, X right, double weight) {
         add(left, right, weight);
         return this;
     }
